tests/requirements.txt
requests
paho-mqtt
websocket-client
toml